+++ controller/src/main/java/com/linbit/linstor/dbcp/migration/k8s/crd/Migration_04_v1_17_0_SpaceTrackingV2.java
@@ -1,7 +1,6 @@
 package com.linbit.linstor.dbcp.migration.k8s.crd;
 
 import com.linbit.linstor.ControllerK8sCrdDatabase;
-import com.linbit.linstor.dbdrivers.GeneratedDatabaseTables;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_17_0;
 import com.linbit.linstor.dbdrivers.k8s.crd.LinstorCrd;
 import com.linbit.linstor.dbdrivers.k8s.crd.LinstorSpec;
@@ -24,20 +23,22 @@ public class Migration_04_v1_17_0_SpaceTrackingV2 extends BaseK8sCrdMigration
     public MigrationResult migrateImpl(ControllerK8sCrdDatabase k8sDbRef) throws Exception
     {
         HashMap<String, LinstorCrd<LinstorSpec<?, ?>>> crdMap;
-        crdMap = (HashMap<String, LinstorCrd<LinstorSpec<?, ?>>>) txFrom.getCrd(GeneratedDatabaseTables.SPACE_HISTORY);
+        crdMap = (HashMap<String, LinstorCrd<LinstorSpec<?, ?>>>) txFrom.getCrd(
+            GenCrdV1_17_0.GeneratedDatabaseTables.SPACE_HISTORY
+        );
         for (LinstorCrd<LinstorSpec<?, ?>> value : crdMap.values())
         {
-            txTo.delete(GeneratedDatabaseTables.SPACE_HISTORY, value);
+            txTo.delete(GenCrdV1_17_0.GeneratedDatabaseTables.SPACE_HISTORY, value);
         }
 
         crdMap.clear();
 
         crdMap = (HashMap<String, LinstorCrd<LinstorSpec<?, ?>>>) txTo.getCrd(
-            GeneratedDatabaseTables.SATELLITES_CAPACITY
+            GenCrdV1_17_0.GeneratedDatabaseTables.SATELLITES_CAPACITY
         );
         for (LinstorCrd<LinstorSpec<?, ?>> value : crdMap.values())
         {
-            txTo.delete(GeneratedDatabaseTables.SATELLITES_CAPACITY, value);
+            txTo.delete(GenCrdV1_17_0.GeneratedDatabaseTables.SATELLITES_CAPACITY, value);
         }
 
         MigrationResult result = new MigrationResult();