+++ controller/src/main/java/com/linbit/linstor/dbcp/migration/k8s/crd/Migration_09_v1_19_1_RandomizeLocalClusterId.java
@@ -1,7 +1,6 @@
 package com.linbit.linstor.dbcp.migration.k8s.crd;
 
 import com.linbit.linstor.ControllerK8sCrdDatabase;
-import com.linbit.linstor.dbdrivers.GeneratedDatabaseTables;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_19_1;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_19_1.PropsContainers;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_19_1.PropsContainersSpec;
@@ -24,7 +23,7 @@ public class Migration_09_v1_19_1_RandomizeLocalClusterId extends BaseK8sCrdMigr
     public MigrationResult migrateImpl(ControllerK8sCrdDatabase k8sDbRef) throws Exception
     {
         List<PropsContainers> list = txFrom.<PropsContainers, PropsContainersSpec>getClient(
-            GeneratedDatabaseTables.PROPS_CONTAINERS
+            GenCrdV1_19_1.GeneratedDatabaseTables.PROPS_CONTAINERS
         ).list();
         String newRandomUuid = UUID.randomUUID().toString().toLowerCase();
         for (PropsContainers propCrd : list)
@@ -33,7 +32,7 @@ public class Migration_09_v1_19_1_RandomizeLocalClusterId extends BaseK8sCrdMigr
             if (propSpec.propKey.equals("Cluster/LocalID"))
             {
                 txTo.replace(
-                    GeneratedDatabaseTables.PROPS_CONTAINERS,
+                    GenCrdV1_19_1.GeneratedDatabaseTables.PROPS_CONTAINERS,
                     GenCrdV1_19_1.createPropsContainers(
                         propSpec.propsInstance,
                         propSpec.propKey,