+++ controller/src/main/java/com/linbit/linstor/dbcp/migration/k8s/crd/Migration_15_v1_19_1_AddEmptySnapNameToLRI.java
@@ -1,7 +1,6 @@
 package com.linbit.linstor.dbcp.migration.k8s.crd;
 
 import com.linbit.linstor.ControllerK8sCrdDatabase;
-import com.linbit.linstor.dbdrivers.GeneratedDatabaseTables;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_19_1;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_19_1.LayerDrbdResourceDefinitions;
 import com.linbit.linstor.dbdrivers.k8s.crd.GenCrdV1_19_1.LayerDrbdResourceDefinitionsSpec;
@@ -29,14 +28,14 @@ public class Migration_15_v1_19_1_AddEmptySnapNameToLRI extends BaseK8sCrdMigrat
     public MigrationResult migrateImpl(ControllerK8sCrdDatabase k8sDbRef) throws Exception
     {
         Collection<LayerResourceIds> crdList = txFrom.<LayerResourceIds, LayerResourceIdsSpec>getCrd(
-            GeneratedDatabaseTables.LAYER_RESOURCE_IDS
+            GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_RESOURCE_IDS
         ).values();
 
         for (LayerResourceIds crd : crdList)
         {
             LayerResourceIdsSpec spec = crd.getSpec();
             txTo.upsert(
-                GeneratedDatabaseTables.LAYER_RESOURCE_IDS,
+                GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_RESOURCE_IDS,
                 GenCrdV1_19_1.createLayerResourceIds(
                     spec.layerResourceId,
                     spec.nodeName,
@@ -52,16 +51,16 @@ public class Migration_15_v1_19_1_AddEmptySnapNameToLRI extends BaseK8sCrdMigrat
 
         Collection<LayerDrbdResourceDefinitions> drbdRscDfnList = txFrom
             .<LayerDrbdResourceDefinitions, LayerDrbdResourceDefinitionsSpec>getCrd(
-                GeneratedDatabaseTables.LAYER_DRBD_RESOURCE_DEFINITIONS
+                GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_DRBD_RESOURCE_DEFINITIONS
             )
             .values();
         for (LayerDrbdResourceDefinitions crd : drbdRscDfnList)
         {
             LayerDrbdResourceDefinitionsSpec spec = crd.getSpec();
             // upsert will not find the old entry since we (might) change the snapshot name.
-            txTo.delete(GeneratedDatabaseTables.LAYER_DRBD_RESOURCE_DEFINITIONS, crd);
+            txTo.delete(GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_DRBD_RESOURCE_DEFINITIONS, crd);
             txTo.create(
-                GeneratedDatabaseTables.LAYER_DRBD_RESOURCE_DEFINITIONS,
+                GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_DRBD_RESOURCE_DEFINITIONS,
                 GenCrdV1_19_1.createLayerDrbdResourceDefinitions(
                     spec.resourceName,
                     spec.resourceNameSuffix,
@@ -79,16 +78,16 @@ public class Migration_15_v1_19_1_AddEmptySnapNameToLRI extends BaseK8sCrdMigrat
 
         Collection<LayerDrbdVolumeDefinitions> drbdVlmDfnList = txFrom
             .<LayerDrbdVolumeDefinitions, LayerDrbdVolumeDefinitionsSpec>getCrd(
-                GeneratedDatabaseTables.LAYER_DRBD_VOLUME_DEFINITIONS
+                GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_DRBD_VOLUME_DEFINITIONS
             )
             .values();
         for (LayerDrbdVolumeDefinitions crd : drbdVlmDfnList)
         {
             LayerDrbdVolumeDefinitionsSpec spec = crd.getSpec();
             // upsert will not find the old entry since we (might) change the snapshot name.
-            txTo.delete(GeneratedDatabaseTables.LAYER_DRBD_VOLUME_DEFINITIONS, crd);
+            txTo.delete(GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_DRBD_VOLUME_DEFINITIONS, crd);
             txTo.create(
-                GeneratedDatabaseTables.LAYER_DRBD_VOLUME_DEFINITIONS,
+                GenCrdV1_19_1.GeneratedDatabaseTables.LAYER_DRBD_VOLUME_DEFINITIONS,
                 GenCrdV1_19_1.createLayerDrbdVolumeDefinitions(
                     spec.resourceName,
                     spec.resourceNameSuffix,